backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdf2image==1.16.3
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.5.0
matplotlib>=3.7.0
python-dotenv>=1.0.0 